+++ tests/testapp/models.py
@@ -1,3 +1,5 @@
+import uuid
+
 from django.db import models
 
 from tree_queries.models import TreeNode
@@ -63,3 +65,13 @@ class AlwaysTreeQueryModel(TreeNode):
 
     class Meta:
         base_manager_name = "objects"
+
+
+class UUIDModel(TreeNode):
+    id = models.UUIDField(primary_key=True, default=uuid.uuid4)
+    name = models.CharField(max_length=100)
+
+    objects = TreeQuerySet.as_manager(with_tree_fields=True)
+
+    def __str__(self):
+        return self.name